GENERALIZE FOR STARTING JAMOS
Fix composeSyllable function in order to check if there is a final jamo.
If there is no final jamo, then set finalIndex to 0.
This prevents the previous problem of finalIndex being set to -1 if there was no final jamo, which thus causes an offf-by-one (-1) error in the Unicode lookup (char before the one it should be).